% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpdd_datasource}
\alias{gpdd_datasource}
\title{The data source table}
\description{
Information on where the data was obtained from, relating to the
Main table through a unique DatasourceID.  Sources of data include
published journals, books and unpublished datasets and the references
details are held here.  The table also contains information regarding
access restrictions, contact details and in what medium the data was obtained.
}
\author{
GPDD Administrator \email{cpb-gpdd-dl@imperial.ac.uk}
}
\references{
\url{http://www3.imperial.ac.uk/cpb/databases/gpdd}
}
\keyword{data}

