% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpdd_main}
\alias{gpdd_main}
\title{main table: metadata for each time series}
\description{
A MAIN record is a 'time series' which is unique
Taxon/Location/LifeCycle combination. Sequential data for multiple
life stages (e.g. eggs, larve and adults) are split into different
Main records and must be amalgamated to create a single time series.
Where more than one adult generation occurs per year generation is
identified in the generation column of the data table
}
\author{
GPDD Administrator \email{cpb-gpdd-dl@imperial.ac.uk}
}
\references{
\url{http://www3.imperial.ac.uk/cpb/databases/gpdd}
}
\keyword{data}

