% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpdd_restricted}
\alias{gpdd_restricted}
\title{Restricted data sets table}
\description{
Due to licensing restrictions 686 series from 6 sources cannot
be distributed without the permission of the owner. These are
data from the British Trust for Ornithology’s Common Bird Census
(97) and Constant Effort Recording Scheme (32), Rothamstead
Experimental Station, UK (9), the National monitoring programme
for wintering wildfowl in Norway 1980 – 93 (T. Nygard, 23),
Phalacrocorax carbo (Great cormorant) and Somateria mollissima
(Common eider) series supplied by N. Rov (2) and data from insect
light trapping supplied by H. Wolda (523).
}
\author{
GPDD Administrator \email{cpb-gpdd-dl@imperial.ac.uk}
}
\references{
\url{http://www3.imperial.ac.uk/cpb/databases/gpdd}
}
\keyword{data}

