% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpdd_taxon}
\alias{gpdd_taxon}
\title{The taxon table}
\description{
The taxon table stores the taxonomic names relating to Main records.
It is links to the MAIN table with a unique TaxonID.  Most series
are for species.  Some extra information regarding breeding habitats
etc may be found in the notes column
}
\author{
GPDD Administrator \email{cpb-gpdd-dl@imperial.ac.uk}
}
\references{
\url{http://www3.imperial.ac.uk/cpb/databases/gpdd}
}
\keyword{data}

