% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpdd_timeperiod}
\alias{gpdd_timeperiod}
\title{The time period table}
\description{
TimePeriod is a look-up table that provides text descriptions of
the temporal period the sample relates to such as 'January',
'Spring', 'Week 1' and 'Day 1'.
}
\author{
GPDD Administrator \email{cpb-gpdd-dl@imperial.ac.uk}
}
\references{
\url{http://www3.imperial.ac.uk/cpb/databases/gpdd}
}
\keyword{data}

