% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpdd_version}
\alias{gpdd_version}
\title{GPDD version information}
\description{
This is Version 2.0 (released 2010).
Version 1.0 (released in 1999) has now been superseded by v2.0 which includes the following enhancements,
}
\details{
-  A consistent definition of a time-series.
- Consistent metadata.
- Units.
- Sampling protocol.
- Consistent temporal coding.
- Addition of missing location information, the spatial bounds of study areas and a spatial accuracy index.
- Abundance data are supplied ‘retro-transformed’ as well as in the published source units.
- Improved documentation.
- 123 additional time-series are included, courtesy of Barry Brook (University of Adelaide).
- Removal of un-cited associated data including body size and biotope information.
}
\author{
GPDD Administrator \email{cpb-gpdd-dl@imperial.ac.uk}
}
\references{
\url{http://www3.imperial.ac.uk/cpb/databases/gpdd}
}
\keyword{data}

